add DENY_ADMIN_CHANGE_ADMIN option

+++ src/users/users.service.ts
@@ -1,4 +1,9 @@
-import { HttpException, HttpStatus, Injectable } from '@nestjs/common';
+import {
+  ForbiddenException,
+  HttpException,
+  HttpStatus,
+  Injectable,
+} from '@nestjs/common';
 import {
   GetAllParam,
   RegisterUser,
@@ -7,13 +12,14 @@ import {
   UserList,
   UserRoles,
   UserUpdate,
-} from './users.dto';
+} from 'src/users/users.dto';
 import { InjectModel } from '@nestjs/sequelize';
 import { UsersModel } from 'src/db/models/users.model';
 import { Op } from 'sequelize';
 import { SessionsModel } from 'src/db/models/sessions.model';
 import { getPasswordHash } from 'src/utils/jsutils.cjs';
-import { errorMessage } from '../utils/error';
+import { errorMessage } from 'src/utils/error';
+import { DENY_ADMIN_CHANGE_ADMIN } from 'src/utils/const';
 
 @Injectable()
 export class UsersService {
@@ -58,14 +64,43 @@ export class UsersService {
       count: data.count,
     };
   }
+  async getUser(login: string): Promise<UsersModel> {
+    const userInfo = await this.userRepository.findOne({
+      where: { login },
+      raw: true,
+    });
+    if (!userInfo)
+      throw new HttpException(
+        errorMessage.UserNotFound,
+        HttpStatus.BAD_REQUEST,
+      );
+    return userInfo;
+  }
   async delete(login: string): Promise<number> {
+    const user = await this.getUser(login);
+    if (+DENY_ADMIN_CHANGE_ADMIN && user.role === UserRoles.admin) {
+      throw new ForbiddenException();
+    }
     await this.sessionRepository.destroy({ where: { login } });
     return this.userRepository.destroy({ where: { login } });
   }
   async update(
     login: string,
     { password, ...param }: UserUpdate,
+    tokenLogin: string,
   ): Promise<[number]> {
+    const userInfo = await this.getUser(login);
+    if (+DENY_ADMIN_CHANGE_ADMIN && userInfo.role === UserRoles.admin) {
+      const blockChangeLock =
+        param.locked !== undefined ? userInfo.locked !== param.locked : false;
+      const byMyself = tokenLogin === login;
+      const blockChangePassword =
+        password !== undefined && !byMyself
+          ? userInfo.password !== getPasswordHash(password)
+          : false;
+      if (blockChangeLock || blockChangePassword)
+        throw new ForbiddenException();
+    }
     return this.userRepository.update(
       {
         ...param,
